requirements.txt
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
streamlit>=1.30.0
plotly>=5.18.0
groq>=0.4.0
python-dotenv>=1.0.0
networkx>=3.0
shap>=0.44.0
fpdf2>=2.7.0
